+++ src/pages/mySchedule/MySchedule.scss
@@ -9,22 +9,4 @@
         font-weight: 400;
         margin-bottom: 4vh;
     }
-
-    button[class*='card'] {
-        margin-top: 2vh;
-        width: 60%;
-        border-radius: 5vw;
-    }
-
-    button.my-card {
-        text-align: right;
-        float: right;
-        border-bottom-right-radius: 0;
-    }
-
-    button.other-card {
-        text-align: left;
-        float: left;
-        border-bottom-left-radius: 0;
-    }
 }
